fixes and other option for guessing coaster

+++ src/Main.java
@@ -2,56 +2,6 @@ import java.util.Scanner;
 
 public class Main {
     private static String APIkey;
-    public static void updateCoasterDB() {
-        APIService service =
-                new APIService(APIkey);
-        service.saveCoastersToFile(service.getAllCoasters(), "coasters.json");
-    }
-
-    public static void testAPI(){
-        APIService service =
-                new APIService(APIkey);
-        service.getFirstCoasters();
-    }
-
-    public static CoasterDB.Order getOrderFromChar(char c) {
-        CoasterDB.Order o;
-        switch(c) {
-            case 't' -> o = CoasterDB.Order.EQUAL;
-            case 'f' -> o = CoasterDB.Order.NOT_EQUAL;
-            case 'g' -> o = CoasterDB.Order.GREATER_THAN;
-            case 'l' -> o = CoasterDB.Order.LESS_THAN;
-            default -> throw new RuntimeException("Unhandled order: " + c);
-        }
-        return o;
-    }
-
-    private static String readValidAnswer(Scanner sc) {
-        while (true) {
-            String input = sc.nextLine().trim().toLowerCase();
-
-            if (input.length() != 7) {
-                System.out.println("Input must have exactly 7 characters!");
-                continue;
-            }
-
-            boolean valid = true;
-            for (char c : input.toCharArray()) {
-                if (c != 't' && c != 'f' && c != 'g' && c != 'l') {
-                    valid = false;
-                    break;
-                }
-            }
-
-            if (!valid) {
-                System.out.println("Only t, f, g and l are allowed!");
-                continue;
-            }
-
-            return input;
-        }
-    }
-
     public static void main(String[] args) {
         if(args.length != 1) {
             System.out.println("Usage: API-Key must be first argument");
@@ -59,13 +9,13 @@ public class Main {
         }
         APIkey = args[0];
 
-
         // ------------------------SETTINGS-------------------------
+        final guessedCoasterChooser chooser = guessedCoasterChooser.TOP_RATED;
         final boolean outputRemovedCoasterInfo = false;
-        final boolean chooseTopRatedCoasterAsGuess = false;
         boolean removeUnrankedCoasters = false;
+        final boolean removeInclompleteCoasters = true;
 
-        if(chooseTopRatedCoasterAsGuess) removeUnrankedCoasters = true;
+        if(chooser== guessedCoasterChooser.TOP_RATED) removeUnrankedCoasters = true;
         // ---------------------------------------------------------
 
         final Scanner sc = new Scanner(System.in);
@@ -74,19 +24,11 @@ public class Main {
 
 
         // remove bad coasters
-         db.removeUncompleteCoasters();
-
+        if(removeInclompleteCoasters) db.removeUncompleteCoasters();
         if(removeUnrankedCoasters) db.removeUnrankedCoasters();
 
-        System.out.println("Average Height of all Coasters: " + db.getAverageHeight());
-        System.out.println("Average Length of all Coasters: " + db.getAverageLength());
-        System.out.println("Average Speed of all Coasters: " + db.getAverageSpeed());
-        System.out.println("Most Common Manufacturer: " + db.getMostCommonManufacturer());
-        System.out.println("Most Common Country: " + db.getMostCommonCountry());
-        System.out.println("Most Common Seating: " + db.getMostCommonSeating());
-
         // Choose Diamondback as first coaster
-        Coaster curCoaster = db.findCoaster("Diamonback");
+        Coaster curCoaster = db.findCoaster("Diamondback");
 
         CoasterDB.Order countryOrder;
         CoasterDB.Order manufacturerOrder;
@@ -118,11 +60,17 @@ public class Main {
             db.keepHeight(heightOrder, curCoaster.height);
             db.keepLength(lengthOrder, curCoaster.length);
             db.keepSpeed(speedOrder, curCoaster.speed);
-            if(chooseTopRatedCoasterAsGuess) curCoaster = db.getTopCoasters(1).getFirst();
+            if(guess==3) {
+                System.out.print("what is the first letter of the coaster?: ");
+                db.keepStartingChar(CoasterDB.Order.EQUAL, sc.nextLine().charAt(0));
+            }
+            if (chooser == guessedCoasterChooser.TOP_RATED) curCoaster = db.getTopCoasters(1).getFirst();
+            else if (chooser == guessedCoasterChooser.RANDOM) curCoaster = db.randomCoaster();
+            else if (chooser == guessedCoasterChooser.AVERAGE_COASTER) curCoaster = curCoaster = db.findMostAverageCoaster();
             else curCoaster = db.randomCoaster();
 
             if(db.coasters.size() == 1) {
-                System.out.println("Coaster found!");
+                System.out.println("Coaster found with "+ guess + " Guesses!");
                 System.out.println(curCoaster);
                 System.exit(0);
             }
@@ -130,5 +78,73 @@ public class Main {
             System.out.println("Reduced Options to: "+ db.coasters.size());
             System.out.println("Guess: " + curCoaster.name + " at " + curCoaster.park);
         }
+    }
+
+    public enum guessedCoasterChooser {
+        TOP_RATED,
+        RANDOM,
+        AVERAGE_COASTER
+    }
+
+    public static void updateCoasterDB() {
+        APIService service =
+                new APIService(APIkey);
+        service.saveCoastersToFile(service.getAllCoasters(), "coasters.json");
+    }
+
+    public static void testAPI(){
+        APIService service =
+                new APIService(APIkey);
+        service.getFirstCoasters();
+    }
+
+    public static CoasterDB.Order getOrderFromChar(char c) {
+        CoasterDB.Order o;
+        switch(c) {
+            case 't' -> o = CoasterDB.Order.EQUAL;
+            case 'f' -> o = CoasterDB.Order.NOT_EQUAL;
+            case 'g' -> o = CoasterDB.Order.GREATER_THAN;
+            case 'l' -> o = CoasterDB.Order.LESS_THAN;
+            default -> throw new RuntimeException("Unhandled order: " + c);
+        }
+        return o;
+    }
+
+    private static String readValidAnswer(Scanner sc) {
+        while (true) {
+            System.out.print("Coastle Response: ");
+            String input = sc.nextLine().trim().toLowerCase();
+
+            if (input.length() != 7) {
+                System.out.println("Input must have exactly 7 characters!");
+                continue;
+            }
+
+            boolean valid = true;
+            for (char c : input.toCharArray()) {
+                if (c != 't' && c != 'f' && c != 'g' && c != 'l') {
+                    valid = false;
+                    break;
+                }
+            }
+
+            if (!valid) {
+                System.out.println("Only t, f, g and l are allowed!");
+                continue;
+            }
+
+            return input;
+        }
+    }
+
+    private static  void randomStuff(CoasterDB db) {
+        System.out.println("Average Height of all Coasters: " + db.getAverageHeight());
+        System.out.println("Average Length of all Coasters: " + db.getAverageLength());
+        System.out.println("Average Speed of all Coasters: " + db.getAverageSpeed());
+        System.out.println("Average Inversions of all Coasters: " + db.getAverageInversions());
+        System.out.println("Most Common Manufacturer: " + db.getMostCommonManufacturer());
+        System.out.println("Most Common Country: " + db.getMostCommonCountry());
+        System.out.println("Most Common Seating: " + db.getMostCommonSeating());
+        System.out.println("Most Average Coaster:\n" + db.findMostAverageCoaster().getInfo());
     }
 }
